♻️ Refactor tag state
