+++ src/Component/Box/TagBox/index.tsx
@@ -3,7 +3,7 @@ import { tagBoxColorStyle, tagStyle } from './style.css';
 
 function TagBox({ name, color }: ITagBox) {
   return (
-    <li className={`${tagStyle} ${tagBoxColorStyle[color]}`}>
+    <li className={`${tagStyle} ${tagBoxColorStyle[color ?? 'color1']}`}>
       <div>{name}</div>
     </li>
   );
