+++ src/Component/Utils/Dropdown/DropdownElement.tsx
@@ -4,16 +4,14 @@ import { css } from '@emotion/react';
 import { ChangeEvent } from 'react';
 import { INPUT_TYPE } from '../../../constants/input';
 import { IDropdownElement } from '../../../types/util';
-import { resetSearchProblemInput } from '../../../utils/resetSearchProblemInputs';
 
 interface IDropdownComponentProps extends IDropdownElement {
-  handleCheckedTags: (tagId: string, isChecked: boolean) => void;
+  handleCheckedTags: (tagId: string, name: string, isChecked: boolean) => void;
 }
 
 function DropdownElement({ id, name, handleCheckedTags }: IDropdownComponentProps) {
   function checkHandler({ target }: ChangeEvent<HTMLInputElement>) {
-    handleCheckedTags(target.id, target.checked);
-    resetSearchProblemInput();
+    handleCheckedTags(target.id, target.name, target.checked);
   }
 
   return (
