+++ src/types/tag.ts
@@ -1,9 +1,10 @@
 export interface ITagBox {
   name: string;
-  color: 'color1' | 'color2' | 'color3' | 'color4';
+  color?: 'color1' | 'color2' | 'color3' | 'color4';
 }
 
 export interface ITagState {
   id: string;
+  name: string;
   isChecked: boolean;
 }
