+++ src/utils/listenOutsideClick.ts
@@ -1,5 +1,5 @@
 import { RefObject } from 'react';
-import { EVENT } from '../../constants';
+import { EVENT } from '../constants';
 
 export default function listenOutsideClick( //menuRef가 가르키는 dom요소(드롭다운)가 아닌 곳이 클릭되면 isOpen false로
   menuRef: RefObject<HTMLDivElement>,
